Finalizacion punto 2
