+++ data/Partidas.json
@@ -33,5 +33,10 @@
         "nombre": "Maxi Final",
         "fecha": "2025-07-04 22:11:27",
         "puntuacion": 375
+    },
+    {
+        "nombre": "Error fix",
+        "fecha": "2025-07-04 22:17:01",
+        "puntuacion": 75
     }
 ]
